Multiple files (#47)
* Various fixes

* Move arel extensions
